Decompile func0f027994

+++ src/include/data.h
@@ -58,7 +58,7 @@ extern u32 var8005ef90;
 extern bool g_ModelDistanceDisabled;
 extern f32 g_ModelDistanceScale;
 extern bool var8005efbc;
-extern u32 var8005efc0;
+extern f32 var8005efc0;
 extern bool (*var8005efc4)(struct model *model, struct modelnode *node);
 extern struct gfxvtx *(*g_ModelVtxAllocatorFunc)(s32 numvertices);
 extern void *var8005efcc;
